requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
sqlalchemy
psycopg2-binary
paho-mqtt
google-cloud-sql